Revert "Sync tabs across all OpenAPI blocks (#2817)"
This reverts commit c31cf5243ab0293ae591078067ad636251d6ef60.
