--- .changeset/fifty-donkeys-talk.md
@@ -1,5 +1,0 @@
----
-'@gitbook/react-openapi': patch
----
-
-Sync tabs across all OpenAPI blocks
